+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Rename `token_auths` database table to `pin_logins`.
 - Redirect to PIN verify page after sending PIN email.
 - Simplify PIN email: remove button, show PIN and login link as plain text.
+- Stay on PIN verify page when PIN is invalid or expired.
 
 ## [6.1.1] - 2026-03-08